解决发布引用js css的路径问题
