+++ config/webpack.common.js
@@ -1,26 +1,9 @@
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const { ProvidePlugin } = require("webpack");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const getEntry = require("./utils");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
-const addEntry = () => {
-  let entryObj = {};
-  getEntry().forEach((item) => {
-    entryObj[item] = {
-      import: path.join(path.dirname(__dirname), "src", item, "index.jsx"),
-      filename: path.join("../dist", item, item) + ".js",
-    };
-  });
-  return entryObj;
-};
-
-const pageExtractCssArray = [];
-getEntry().forEach((item) => {
-  pageExtractCssArray.push(
-    new MiniCssExtractPlugin({ filename: item + "/[name].[contenthash].css" })
-  );
-});
 
 const recursiveIssuer = (m) => {
   if (m.issuer) {
@@ -34,9 +17,9 @@ const recursiveIssuer = (m) => {
 
 const commonConfig = {
   context: path.resolve(path.dirname(__dirname), "src"),
-  entry: addEntry(),
   output: {
-    path: path.resolve(__dirname, "..", "./dist"),
+    path: path.resolve(__dirname, "."),
+    publicPath: ''
   },
   module: {
     rules: [
@@ -61,35 +44,10 @@ const commonConfig = {
   resolve: {
     extensions: ["*", ".js", ".json", ".jsx", ".ts", ".tsx"],
   },
-  optimization: {
-    splitChunks: {
-      chunks: "all",
-      cacheGroups: {
-        vendors: {
-          test: /[\\/]node_modules[\\/]/,
-          priority: -10,
-          name: "common",
-          filename: (e) => {
-            return path.join("../dist", "vendors", "common") + ".js";
-          },
-        },
-        default: {
-          minChunks: 2,
-          priority: -20,
-          reuseExistingChunk: true,
-        },
-      },
-      // cacheGroups: addCssSplitChunk()
-    },
-  },
+  
   plugins: [
     new ProvidePlugin({
       React: "react",
-    }),
-    new MiniCssExtractPlugin({
-      filename: (item) => {
-        return path.join("../dist", item.chunk.name, item.chunk.name) + ".css";
-      },
     }),
   ],
 };
@@ -105,7 +63,7 @@ getEntry().forEach((pathname) => {
       "template",
       "index.html"
     ),
-    publicPath: "/",
+    publicPath: "",
     chunks: ["manifest", "vendor", pathname],
   };
   commonConfig.plugins.push(new HtmlWebpackPlugin(conf));
